+++ prelogout.php
@@ -1,29 +1,29 @@
 <?php
 /**
- * MumieTaskHook plugin
+ * MumieTaskHook plugin.
  *
  * @copyright   2019 integral-learning GmbH (https://www.integral-learning.de/)
  * @author      Tobias Goltz (tobias.goltz@integral-learning.de)
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-
-$logouturls = urldecode($_GET["logoutUrl"]);
-$redirect = json_encode($_GET["redirect"]);
+$logouturls = $_GET['logoutUrl'];
+$redirect = json_encode($_GET['redirect']);
 ?>
 
 <script>
-    var logouturls = Object.values(<?php echo $logouturls ?>);
-    var promises = [];
-    logouturls.forEach(function(url) {
+    const logouturls = Object.values(<?php echo $logouturls; ?>);
+    const promises = [];
+    logouturls.forEach(function (url) {
         promises.push(logoutFromServer(url));
     });
     Promise.all(promises)
-    .then(function (values) {
-        window.location.href =<?php echo $redirect ?>;
-    });
+        .then(function () {
+            window.location.href =<?php echo $redirect; ?>;
+        });
+
     function logoutFromServer(url) {
-        var promise = new Promise(function (resolve, reject) {
-            var request = new XMLHttpRequest();
+        return new Promise(function (resolve) {
+            const request = new XMLHttpRequest();
             request.open("GET", url);
             request.withCredentials = true;
             request.timeout = 10000;
@@ -31,10 +31,9 @@ $redirect = json_encode($_GET["redirect"]);
             request.onreadystatechange = function () {
                 resolve();
             }
-            request.ontimeout = function (e) {
+            request.ontimeout = function () {
                 resolve();
             }
-        })
-        return promise;
+        });
     }
 </script>